Setting up contentTypeMenuBar - Assets required

+++ IGN Code Foo 2018 - App/IGN Code Foo 2018 - App/Helpers/Extentions.swift
@@ -21,11 +21,11 @@ extension UIView {
 
 
 extension UIColor {
-    convenience init(red: Int, green: Int, blue: Int) {
+    convenience init(red: CGFloat, green: CGFloat, blue: CGFloat) {
         assert(red >= 0 && red <= 255, "Invalid red component")
         assert(green >= 0 && green <= 255, "Invalid green component")
         assert(blue >= 0 && blue <= 255, "Invalid blue component")
         
-        self.init(red: CGFloat(red) / 255.0, green: CGFloat(green) / 255.0, blue: CGFloat(blue) / 255.0, alpha: 1.0)
+        self.init(red: red/255, green: green/255, blue: blue/255, alpha: 1)
     }
 }
